Store language defaults to browser language
Now the scripts tries to detect the browser language and use it as the store language default. There is a fallback to 'us' if for whatever reason the browser does not give a language. I couldn't replicate the dependencies stack, so the source file should be re-compiled by someone with the full-stack.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>